Rename lib to src, use relative imports
