+++ lambda/src/email.py
