+++ lambda/src/exceptions.py
