+++ lambda/src/handlers/check_in.py
@@ -1,7 +1,7 @@
 import logging
 import sys
 
-from lib import swa, exceptions  # NOQA
+from .. import swa, exceptions
 
 # Set up logging
 log = logging.getLogger(__name__)
