+++ lambda/src/swa.py
